Перенес метод convert в MovieQuizPresenter и обновил MovieQuizViewController; тесты проходят
